3c307a5a492762358fe6de3ebe0caa8c/external_parse_data/requirements.txt
beautifulsoup4==4.11.1
psycopg2-binary==2.9.3
selenium==4.4.3
webdriver-manager==3.8.3